Remove default values for port and containerPort
Having default values for `port` and `containerPort` is preventing the possibility of using `ports`, introduced in my previous PR.

Even though this is a breaking change for deployments that have not set a value for `port` and `containerPort`, I haven't seen that any of us have actually skipped declaring them. Nevertheless, deployments are using a fixed chart version, so making this change should not be an issue at all.